+++ src/main/java/zmaster587/libVulpes/Configuration.java
@@ -4,5 +4,5 @@ public class Configuration {
 
 	public static double EUMult;
 	public static float powerMult;
-
+	public static boolean defaultMultiblockMachineEnabled = true;
 }