+++ src/main/java/zmaster587/libVulpes/tile/multiblock/TileMultiblockMachine.java
@@ -10,12 +10,12 @@ import net.minecraft.nbt.NBTTagList;
 import net.minecraft.network.NetworkManager;
 import net.minecraft.network.play.server.SPacketUpdateTileEntity;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.SoundEvent;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.oredict.OreDictionary;
+import zmaster587.libVulpes.Configuration;
 import zmaster587.libVulpes.interfaces.IRecipe;
 import zmaster587.libVulpes.network.PacketHandler;
 import zmaster587.libVulpes.network.PacketMachine;
@@ -44,6 +44,7 @@ public abstract class TileMultiblockMachine extends TileMultiPowerConsumer {
 
 	public TileMultiblockMachine() {
 		super();
+		enabled = Configuration.defaultMultiblockMachineEnabled;
 		outputItemStacks = null;
 	}
 
@@ -172,6 +173,11 @@ public abstract class TileMultiblockMachine extends TileMultiPowerConsumer {
 		outputItemStacks = null;
 		outputFluidStacks = null;
 		super.deconstructMultiBlock(world, destroyedPos, blockBroken, state);
+		enabled = Configuration.defaultMultiblockMachineEnabled;
+		if (!world.isRemote) {
+			markDirty();
+			world.notifyBlockUpdate(pos, world.getBlockState(pos), world.getBlockState(pos), 3);
+		}
 	}
 
 	protected void processComplete() {